+++ client/src/App.js
@@ -2,14 +2,14 @@ import React, { Fragment } from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 
 import { Navbar } from "./components/layout/Navbar";
-import  Alert from "./components/layout/Alert";
-import { Login } from "./components/auth/Login";
-import  Register  from "./components/auth/Register";
+import Alert from "./components/layout/Alert";
+import Login from "./components/auth/Login";
+import Register from "./components/auth/Register";
 import Dashboard from "./components/dashboard/Dashboard";
 import TicketsTable from "./components/ticket/TicketTable";
 
 // Redux
-import { Provider } from 'react-redux';
+import { Provider } from "react-redux";
 import store from "./store";
 
 import "./App.css";
