+++ client/src/reducers/auth.js
@@ -1,4 +1,4 @@
-import { REGISTER_SUCCESS, REGISTER_FAIL } from "../actions/types";
+import { REGISTER_SUCCESS, REGISTER_FAIL, LOGIN_SUCCESS } from "../actions/types";
 
 const initialState = {
   token: localStorage.getItem("token"),
@@ -12,6 +12,7 @@ export default function (state = initialState, action) {
 
   switch (type) {
     case REGISTER_SUCCESS:
+    case LOGIN_SUCCESS:
       localStorage.setItem("token", payload.token);
       return {
         ...state,
